Move `Paths` and `NymConfigPaths` into `nym-vpn-lib` and tidy-up implementation.
